+++ vgui2-dev/include/IInputInternal.h
