+++ vgui2-dev/include/ILocalize.h
