+++ vgui_support/wscript
@@ -98,29 +98,6 @@ def configure(conf):
 		except conf.errors.ConfigurationError:
 			conf.fatal("Can't compile simple program. Check your path to vgui-dev repository.")
 
-def build_vgui(bld):
-	libs = []
-
-	# basic build: dedicated only, no dependencies
-	if bld.env.DEST_OS != 'win32':
-		libs += ['DL','M']
-
-	libs.append('VGUI')
-
-	source = bld.path.ant_glob(['*.cpp', 'vgui2/*.cpp', 'vinterface/*.cpp'])
-
-	includes = [ '.', 'vinterface', 'filesystem', '../common', '../engine', '../public' ]
-
-	bld.shlib(
-		source   = source,
-		target   = 'vgui_support',
-		features = 'cxx',
-		includes = includes,
-		use      = libs,
-		install_path = bld.env.LIBDIR,
-		subsystem = bld.env.MSVC_SUBSYSTEM
-	)
-
 def build_fs(bld):
 	libs = []
 
@@ -128,9 +105,10 @@ def build_fs(bld):
 	if bld.env.DEST_OS != 'win32':
 		libs += ['DL','M']
 
-	source = bld.path.ant_glob(['vinterface/*.cpp', 'filesystem/*.cpp'])
+	source = bld.path.ant_glob(['filesystem/*.cpp'])
+	source += bld.path.parent.ant_glob(['vgui2-dev/src/interface.cpp'])
 
-	includes = [ '.', 'vinterface', 'filesystem', '../common', '../engine', '../public' ]
+	includes = [ '.', '../vgui2-dev/include', '../common', '../engine', '../public' ]
 
 	bld.shlib(
 		source   = source,
@@ -142,9 +120,33 @@ def build_fs(bld):
 		subsystem = bld.env.MSVC_SUBSYSTEM
 	)
 
+def build_vgui(bld):
+	libs = []
+
+	# basic build: dedicated only, no dependencies
+	if bld.env.DEST_OS != 'win32':
+		libs += ['DL','M']
+
+	libs.append('VGUI')
+
+	source = bld.path.ant_glob(['*.cpp'])
+	source += bld.path.parent.ant_glob(['vgui2-dev/src/*.cpp'])
+
+	includes = [ '.', '../vgui2-dev/include', '../common', '../engine', '../public' ]
+
+	bld.shlib(
+		source   = source,
+		target   = 'vgui_support',
+		features = 'cxx',
+		includes = includes,
+		use      = libs,
+		install_path = bld.env.LIBDIR,
+		subsystem = bld.env.MSVC_SUBSYSTEM
+	)
+
 def build(bld):
 	if bld.env.NO_VGUI:
 		return
 
-	build_vgui(bld)
 	build_fs(bld)
+	build_vgui(bld)
